move case classes from router model
