+++ src/main/scala/beam/agentsim/agents/choice/mode/ModeChoiceRideHailIfAvailable.scala
@@ -3,7 +3,7 @@ package beam.agentsim.agents.choice.mode
 import beam.agentsim.agents.modalbehaviors.ModeChoiceCalculator
 import beam.router.Modes
 import beam.router.Modes.BeamMode.RIDE_HAIL
-import beam.router.RoutingModel.EmbodiedBeamTrip
+import beam.router.model.EmbodiedBeamTrip
 import beam.sim.BeamServices
 
 /**
