+++ src/main/scala/beam/agentsim/agents/choice/mode/ModeChoiceTransitIfAvailable.scala
@@ -2,7 +2,7 @@ package beam.agentsim.agents.choice.mode
 
 import beam.agentsim.agents.modalbehaviors.ModeChoiceCalculator
 import beam.router.Modes
-import beam.router.RoutingModel.EmbodiedBeamTrip
+import beam.router.model.EmbodiedBeamTrip
 import beam.sim.BeamServices
 
 /**
